+++ components/Layout.js
@@ -19,7 +19,6 @@ class Layout extends React.PureComponent {
         <Head>
           <script src="ace/src-min/ace.js" />
           <link rel="stylesheet" type="text/css" href="markup-editor/lib/codemirror.css" />
-          <link href="https://fonts.googleapis.com/css?family=Roboto|Roboto+Mono" rel="stylesheet" />
           <link rel="stylesheet" type="text/css" href="hljs/styles/github.css" />
           <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/github-markdown-css/2.9.0/github-markdown.min.css" />
         </Head>