+++ src/components/outline/SortableList.js
@@ -12,17 +12,17 @@ const SortableItem = SortableElement(({
   const key = `${value.content}${value.index}`;
   return (
     <li className={`item level-${value.level}`}>
-      <button
-        className={`toggle ${value.children.length > 0 ? '' : 'invisible'} ${hidden[key] ? 'closed' : 'opened'}`}
-        onClick={() => onSectionToggle(key)}
-      />
       <button className="item-wrapper" onClick={() => onItemClick(value)}>
-        <DragHandle />
+        <button
+          className={`toggle ${value.children.length > 0 ? '' : 'invisible'} ${hidden[key] ? 'closed' : 'opened'}`}
+          onClick={() => onSectionToggle(key)}
+        />
         <span
           className="outlineItem"
           // eslint-disable-next-line react/no-danger
           dangerouslySetInnerHTML={{ __html: value.content.replace(/<[^>]*>/g, '') }}
         />
+        <DragHandle />
       </button>
       {!hidden[key] &&
         <SortableList
@@ -32,6 +32,7 @@ const SortableItem = SortableElement(({
           onOrderChange={onOrderChange}
           onItemClick={onItemClick}
           onSectionToggle={onSectionToggle}
+          useDragHandle
         />
       }
     </li>